+++ app/src/main/java/org/nutritionfacts/dailydozen/Args.java
@@ -6,6 +6,7 @@ public class Args {
     public static final String DATES_WITH_PARTIAL_SERVINGS = "dates_with_partial_servings";
     public static final String FOOD_ID = "food_id";
     public static final String TWEAK_ID = "tweak_id";
+    public static final String SCROLL_TO_HISTORY = "scroll_to_history";
     public static final String OPEN_NOTIFICATION_SETTINGS = "open_notification_settings";
     public static final String UPDATE_REMINDER_PREF = "update_reminder_pref";
