+++ app/src/main/java/org/nutritionfacts/dailydozen/Common.java
@@ -17,10 +17,8 @@ import androidx.core.content.ContextCompat;
 
 import com.google.android.material.color.MaterialColors;
 
-import org.nutritionfacts.dailydozen.activity.FoodInfoActivity;
 import org.nutritionfacts.dailydozen.activity.HistoryActivity;
 import org.nutritionfacts.dailydozen.activity.ServingsHistoryActivity;
-import org.nutritionfacts.dailydozen.activity.TweakInfoActivity;
 import org.nutritionfacts.dailydozen.activity.TweakServingsHistoryActivity;
 import org.nutritionfacts.dailydozen.activity.WeightHistoryActivity;
 import org.nutritionfacts.dailydozen.activity.DateSelectionHost;
@@ -147,12 +145,12 @@ public class Common {
         if (isSupplement(food)) {
             openUrlInExternalBrowser(context, FoodInfo.getFoodTypeVideosLink(food.getName()));
         } else {
-            context.startActivity(createFoodIntent(context, FoodInfoActivity.class, food));
+            startSelectableDateActivity(context, createFoodIntent(context, HistoryActivity.class, food));
         }
     }
 
     public static void openTweakInfo(final Context context, final Tweak tweak) {
-        context.startActivity(createTweakIntent(context, TweakInfoActivity.class, tweak));
+        startSelectableDateActivity(context, createTweakIntent(context, HistoryActivity.class, tweak));
     }
 
     public static boolean isSupplement(final Food food) {
@@ -160,11 +158,15 @@ public class Common {
     }
 
     public static void openFoodHistory(final Context context, final Food food) {
-        startSelectableDateActivity(context, createFoodIntent(context, HistoryActivity.class, food));
+        final Intent intent = createFoodIntent(context, HistoryActivity.class, food);
+        intent.putExtra(Args.SCROLL_TO_HISTORY, true);
+        startSelectableDateActivity(context, intent);
     }
 
     public static void openTweakHistory(final Context context, final Tweak tweak) {
-        startSelectableDateActivity(context, createTweakIntent(context, HistoryActivity.class, tweak));
+        final Intent intent = createTweakIntent(context, HistoryActivity.class, tweak);
+        intent.putExtra(Args.SCROLL_TO_HISTORY, true);
+        startSelectableDateActivity(context, intent);
     }
 
     public static void openServingsHistory(final Context context) {
